requirements.txt
streamlit==1.47.1
pandas==2.3.1
pillow==11.3.0
pdf2image==1.17.0
PyMuPDF==1.25.3
streamlit-drawable-canvas==0.8.0